refactor(argocd): break up argo_status_updater.go
The 607-line file bundled three types and a set of pure numeric
utilities. Split it into cohesive same-package files and relocate the
generic helpers, moving all code verbatim:

- deployment_monitor.go  DeploymentMonitor and its rollout-polling
                         methods, retry sentinels, and the free
                         functions only it uses
- git_updater.go         GitUpdater (git write-back orchestration and
                         per-repo locking)
- argo_status_updater.go ArgoStatusUpdater, Init, WaitForRollout,
                         sendNotification

The saturating duration/int helpers (mulDurationSaturating,
ceilDivDuration, safeIntToUint) had no dependency on argocd, so they
move to internal/helpers as exported MulDurationSaturating,
CeilDivDuration, and SafeIntToUint, with their tests. No behavior
change: identical declaration set, call sites updated to helpers.*.
