+++ internal/argocd/argo_status_updater.go
@@ -1,420 +1,20 @@
 package argocd
 
 import (
-	"bytes"
 	"context"
 	"errors"
 	"fmt"
 	"log/slog"
-	"math"
 	"net/http"
-	"strings"
 	"time"
-
-	"github.com/shini4i/argo-watcher/internal/config"
-	"github.com/shini4i/argo-watcher/internal/notifications"
-	"github.com/shini4i/argo-watcher/internal/prometheus"
-
-	"github.com/shini4i/argo-watcher/internal/helpers"
 
 	"github.com/avast/retry-go/v4"
 
+	"github.com/shini4i/argo-watcher/internal/config"
 	"github.com/shini4i/argo-watcher/internal/lock"
 	"github.com/shini4i/argo-watcher/internal/models"
-	"github.com/shini4i/argo-watcher/internal/updater"
+	"github.com/shini4i/argo-watcher/internal/notifications"
 )
-
-const (
-	failedToUpdateTaskStatusTemplate = "Failed to change task status: %s"
-	// legacyRetryIntervals preserves the historical 15-step retry window (assumes 15s retry delay, totaling ~3m45s).
-	legacyRetryIntervals = 15
-)
-
-// errForceRetry is an internal sentinel used to keep retry-go polling while the rollout has not reached a final state.
-// It must never reach the user-visible task status — WaitRollout swallows it so the caller can report the actual rollout state instead.
-var errForceRetry = errors.New("force retry")
-
-// errTaskSuperseded is an internal sentinel returned by the poll loop when the task
-// has been marked cancelled in the shared state by a newer deployment for the same
-// app. Unlike errForceRetry it is not swallowed: WaitForRollout uses it to stop
-// without overwriting the "cancelled" status the newer deployment already wrote.
-var errTaskSuperseded = errors.New("task superseded by a newer deployment")
-
-// DeploymentMonitor encapsulates the logic for tracking ArgoCD application rollouts.
-type DeploymentMonitor struct {
-	argo             Argo
-	registryProxyUrl string
-	retryOptions     []retry.Option
-	acceptSuspended  bool
-	retryDelay       time.Duration
-	defaultAttempts  uint
-	// refreshApp is the instance-wide default for requesting an ArgoCD refresh during status checks.
-	// A per-task Refresh override takes precedence (see resolveRefresh).
-	refreshApp bool
-}
-
-// GitUpdater encapsulates the logic required to update Git repositories watched by ArgoCD.
-type GitUpdater struct {
-	locker        lock.Locker
-	repoCachePath string
-	// metrics records lock-wait and write-back durations. May be nil in tests, in which
-	// case observation is skipped.
-	metrics prometheus.MetricsInterface
-}
-
-// NewDeploymentMonitor creates a deployment monitor with the supplied configuration.
-func NewDeploymentMonitor(argo Argo, registryProxyUrl string, retryOptions []retry.Option, acceptSuspended bool, retryDelay time.Duration) *DeploymentMonitor {
-	return &DeploymentMonitor{
-		argo:             argo,
-		registryProxyUrl: registryProxyUrl,
-		retryOptions:     retryOptions,
-		acceptSuspended:  acceptSuspended,
-		retryDelay:       retryDelay,
-	}
-}
-
-// BeginTracking increments the in-progress task counter.
-func (monitor *DeploymentMonitor) BeginTracking() {
-	monitor.argo.metrics.AddInProgressTask()
-}
-
-// EndTracking decrements the in-progress task counter.
-func (monitor *DeploymentMonitor) EndTracking() {
-	monitor.argo.metrics.RemoveInProgressTask()
-}
-
-// resolveRefresh reports whether this task's status check should request an ArgoCD refresh.
-// A per-task Refresh override wins over the instance-wide default; a nil override (field omitted
-// by the client) keeps the default, so old clients behave exactly as before (issue #334).
-func (monitor *DeploymentMonitor) resolveRefresh(task models.Task) bool {
-	if task.Refresh != nil {
-		return *task.Refresh
-	}
-	return monitor.refreshApp
-}
-
-// FetchApplication retrieves the ArgoCD application by name. The context bounds the underlying
-// API call so callers polling under a deadline are not blocked past it.
-//
-// When a refresh is requested, the call is timed and its duration recorded (argocd_refresh_duration_seconds)
-// so slow or stuck refreshes are diagnosable. A stuck refresh is avoided operationally, not here: set the
-// per-task Refresh override (or ARGO_REFRESH_APP) to false for apps that never settle (issue #334).
-func (monitor *DeploymentMonitor) FetchApplication(ctx context.Context, appName string, refresh bool) (*models.Application, error) {
-	if !refresh {
-		return monitor.argo.api.GetApplication(ctx, appName, false)
-	}
-
-	start := time.Now()
-	app, err := monitor.argo.api.GetApplication(ctx, appName, true)
-	monitor.argo.metrics.ObserveRefreshDuration(appName, time.Since(start).Seconds())
-	return app, err
-}
-
-// StoreInitialAppStatus caches the initial rollout status for comparison during monitoring.
-func (monitor *DeploymentMonitor) StoreInitialAppStatus(task *models.Task, application *models.Application) error {
-	if application == nil {
-		return errors.New("application is nil")
-	}
-
-	status := application.GetRolloutStatus(task.ListImages(), monitor.registryProxyUrl, monitor.acceptSuspended)
-	// The ArgoCD API may return images in different orders between calls; sorting guarantees stable hash comparisons.
-	normalizedImages := helpers.NormalizeImages(application.Status.Summary.Images)
-
-	task.SavedAppStatus = models.SavedAppStatus{
-		Status:     status,
-		ImagesHash: helpers.GenerateHash(strings.Join(normalizedImages, ",")),
-	}
-
-	return nil
-}
-
-// WaitRollout polls the application status until it reaches a final state or times out.
-//
-// The timeout is enforced as a wall-clock deadline, not merely as a fixed number of poll
-// attempts: a context deadline bounds the whole loop (and, through FetchApplication, each
-// individual API call). This prevents the deployment from running far past its configured
-// timeout when ArgoCD responds slowly — the failure mode reported in issue #304.
-//
-// Before each poll it re-reads the task status from the shared state: if a newer
-// deployment for the same app has marked this task "cancelled" (issue #353), it
-// stops immediately so no further ArgoCD API calls are made. Because the check
-// goes through the shared state, this works across replicas in an HA setup — the
-// cancelling deployment may be handled by a different replica than this poller.
-func (monitor *DeploymentMonitor) WaitRollout(task models.Task) (*models.Application, error) {
-	// application holds the most recent successfully-fetched state. It is deliberately assigned only
-	// inside the success branch so that a fetch aborted by the deadline (which returns a nil application)
-	// cannot clobber the last-known-good status we want to report on timeout.
-	var application *models.Application
-
-	refresh := monitor.resolveRefresh(task)
-	retryOptions, deadline := monitor.configureRetryOptions(task)
-
-	ctx, cancel := context.WithTimeout(context.Background(), deadline)
-	defer cancel()
-	retryOptions = append(retryOptions, retry.Context(ctx))
-
-	slog.Debug("Waiting for rollout", "id", task.Id, "deadline", deadline)
-
-	err := retry.Do(func() error {
-		// Stop before hitting ArgoCD if a newer deployment superseded this task.
-		// The check is per-iteration: a cancellation that lands mid-iteration is
-		// caught on the next poll, and if this iteration reaches a final state
-		// first, its terminal status wins over "cancelled". That last-writer race
-		// is accepted (issue #353) to avoid a status-conditional write.
-		if monitor.taskSuperseded(task.Id) {
-			return retry.Unrecoverable(errTaskSuperseded)
-		}
-
-		app, fetchErr := monitor.FetchApplication(ctx, task.App, refresh)
-		if fetchErr != nil {
-			return handleApplicationFetchError(task, fetchErr)
-		}
-		application = app
-
-		// Early return for fire and forget mode
-		if app.IsFireAndForgetModeActive() {
-			slog.Debug("Fire and forget mode is active, skipping checks...", "id", task.Id)
-			return nil
-		}
-
-		return checkRolloutStatus(task, app, monitor.registryProxyUrl, monitor.acceptSuspended)
-	}, retryOptions...)
-
-	// Once the retry budget or the wall-clock deadline is exhausted while still polling, surface the
-	// last successfully-fetched application instead of the internal sentinel or the context error, so
-	// the caller can report the real rollout status (e.g. "not synced", "not healthy") to the user.
-	// If no fetch ever succeeded (application is nil), the error is returned so the caller can classify
-	// the underlying failure (e.g. connection refused -> aborted).
-	if application != nil && (errors.Is(err, errForceRetry) || errors.Is(err, context.DeadlineExceeded) || errors.Is(err, context.Canceled)) {
-		err = nil
-	}
-
-	return application, err
-}
-
-// mulDurationSaturating returns count*unit, clamped to math.MaxInt64 to avoid the uint->int64
-// overflow that a very large (client-supplied) attempt count could otherwise wrap into a negative
-// duration. A non-positive unit or zero count yields 0.
-func mulDurationSaturating(count uint, unit time.Duration) time.Duration {
-	if unit <= 0 || count == 0 {
-		return 0
-	}
-	// unit > 0 is guaranteed above, so the quotient is a non-negative int64 that fits in uint64.
-	maxCount := uint64(math.MaxInt64 / int64(unit)) // #nosec G115 -- non-negative quotient
-	if uint64(count) > maxCount {
-		return time.Duration(math.MaxInt64)
-	}
-	// The guard above proves count fits in a positive int64, so this conversion cannot overflow.
-	return time.Duration(count) * unit // #nosec G115 -- count bounded by check above
-}
-
-// ceilDivDuration returns the ceiling of d/unit as an int64, with a minimum of 1.
-// A non-positive unit is treated as invalid and returns 1 to avoid division by zero.
-func ceilDivDuration(d, unit time.Duration) int64 {
-	if unit <= 0 {
-		return 1
-	}
-	result := int64((d + unit - 1) / unit)
-	if result <= 0 {
-		return 1
-	}
-	return result
-}
-
-// safeIntToUint converts an int64 to uint with overflow protection, enforcing a minimum of 1.
-// On 32-bit platforms where uint is 32 bits, values exceeding math.MaxUint32 are clamped to max uint.
-func safeIntToUint(v int64) uint {
-	if v <= 0 {
-		return 1
-	}
-	maxUint := ^uint(0)
-	if uint64(v) > uint64(maxUint) {
-		return maxUint
-	}
-	return uint(v) // #nosec G115 -- overflow checked above
-}
-
-// configureRetryOptions derives retry attempts by preferring per-task overrides, then monitor defaults, and finally a legacy retry window aligned with the current retry delay.
-//
-// Alongside the retry options it returns the wall-clock deadline for the whole polling loop,
-// computed as attempts*delay. The attempt count still caps the number of polls, while the
-// deadline caps the elapsed time so that slow API responses cannot stretch the loop past the
-// intended timeout (see WaitRollout).
-func (monitor *DeploymentMonitor) configureRetryOptions(task models.Task) ([]retry.Option, time.Duration) {
-	retryOptions := append([]retry.Option{}, monitor.retryOptions...)
-
-	delay := monitor.retryDelay
-	if delay <= 0 {
-		delay = ArgoSyncRetryDelay
-	}
-
-	retryOptions = append(retryOptions, retry.Delay(delay))
-
-	delaySeconds := ceilDivDuration(delay, time.Second)
-
-	defaultAttempts := monitor.defaultAttempts
-	if defaultAttempts == 0 {
-		// Legacy fallback: legacyRetryIntervals steps at ArgoSyncRetryDelay pace, scaled to the current delay.
-		fallbackWindow := time.Duration(legacyRetryIntervals) * ArgoSyncRetryDelay
-		defaultAttempts = safeIntToUint(ceilDivDuration(fallbackWindow, delay))
-	}
-
-	if task.Timeout <= 0 {
-		slog.Debug(fmt.Sprintf("Task timeout is non-positive, defaulting to %d attempts", defaultAttempts), "id", task.Id)
-		return append(retryOptions, retry.Attempts(defaultAttempts)), mulDurationSaturating(defaultAttempts, delay)
-	}
-
-	attempts := safeIntToUint(int64(task.Timeout)/delaySeconds + 1)
-
-	slog.Debug(fmt.Sprintf(
-		"Overriding task timeout to %ds with retry delay %s (~%d second step, %d attempts)",
-		task.Timeout,
-		delay,
-		delaySeconds,
-		attempts,
-	), "id", task.Id)
-
-	return append(retryOptions, retry.Attempts(attempts)), mulDurationSaturating(attempts, delay)
-}
-
-// ProcessDeploymentResult determines if the deployment was successful and updates the appropriate status and metrics.
-func (monitor *DeploymentMonitor) ProcessDeploymentResult(task *models.Task, application *models.Application) {
-	status := application.GetRolloutStatus(task.ListImages(), monitor.registryProxyUrl, monitor.acceptSuspended)
-	if application.IsFireAndForgetModeActive() {
-		status = models.ArgoRolloutAppSuccess
-	}
-
-	if status == models.ArgoRolloutAppSuccess {
-		monitor.handleDeploymentSuccess(task)
-	} else {
-		monitor.handleDeploymentFailure(task, status, application)
-	}
-}
-
-// taskSuperseded reports whether the task has been marked cancelled in the shared
-// state, i.e. a newer deployment for the same app has superseded it. A read error
-// is treated as "not superseded" so a transient state hiccup does not abort an
-// otherwise healthy rollout; the check runs again on the next poll.
-func (monitor *DeploymentMonitor) taskSuperseded(id string) bool {
-	current, err := monitor.argo.State.GetTask(id)
-	if err != nil {
-		slog.Warn(fmt.Sprintf("Could not read task status to check for supersession: %s", err), "id", id)
-		return false
-	}
-	return current.Status == models.StatusCancelledMessage
-}
-
-// HandleArgoAPIFailure processes API errors and updates task status accordingly.
-// task is taken by pointer so the resolved terminal status is reflected back to
-// the caller, keeping the outgoing failure notification in sync with the stored
-// status (mirroring handleDeploymentSuccess/handleDeploymentFailure).
-func (monitor *DeploymentMonitor) HandleArgoAPIFailure(task *models.Task, err error) {
-	monitor.argo.metrics.AddFailedDeployment(task.App)
-	finalStatus := determineFailureStatus(*task, err)
-	reason := fmt.Sprintf(ArgoAPIErrorTemplate, err.Error())
-	slog.Warn(fmt.Sprintf("Deployment failed with status \"%s\". Aborting with error: %s", finalStatus, reason), "id", task.Id)
-
-	if err := monitor.argo.State.SetTaskStatus(task.Id, finalStatus, reason); err != nil {
-		slog.Error(fmt.Sprintf(failedToUpdateTaskStatusTemplate, err), "id", task.Id)
-	}
-	task.Status = finalStatus
-}
-
-func (monitor *DeploymentMonitor) handleDeploymentSuccess(task *models.Task) {
-	slog.Info("App is running on the expected version.", "id", task.Id)
-	monitor.argo.metrics.ResetFailedDeployment(task.App)
-	if err := monitor.argo.State.SetTaskStatus(task.Id, models.StatusDeployedMessage, ""); err != nil {
-		slog.Error(fmt.Sprintf(failedToUpdateTaskStatusTemplate, err), "id", task.Id)
-	}
-	task.Status = models.StatusDeployedMessage
-}
-
-func (monitor *DeploymentMonitor) handleDeploymentFailure(task *models.Task, status string, application *models.Application) {
-	slog.Info("App deployment failed.", "id", task.Id)
-	monitor.argo.metrics.AddFailedDeployment(task.App)
-	reason := fmt.Sprintf(
-		"Application deployment failed. Rollout status is %s\n\n%s",
-		status,
-		application.GetRolloutMessage(status, task.ListImages()),
-	)
-	if err := monitor.argo.State.SetTaskStatus(task.Id, models.StatusFailedMessage, reason); err != nil {
-		slog.Error(fmt.Sprintf(failedToUpdateTaskStatusTemplate, err), "id", task.Id)
-	}
-	task.Status = models.StatusFailedMessage
-}
-
-// NewGitUpdater creates a GitUpdater instance. metrics may be nil, in which case
-// duration observation is skipped.
-func NewGitUpdater(locker lock.Locker, repoCachePath string, metrics prometheus.MetricsInterface) *GitUpdater {
-	return &GitUpdater{
-		locker:        locker,
-		repoCachePath: repoCachePath,
-		metrics:       metrics,
-	}
-}
-
-// observeLockWait records how long a task waited to acquire the per-repo lock.
-func (gitUpdater *GitUpdater) observeLockWait(app string, d time.Duration) {
-	if gitUpdater.metrics != nil {
-		gitUpdater.metrics.ObserveGitLockWaitDuration(app, d.Seconds())
-	}
-}
-
-// observeWriteback records how long the git write-back took while holding the lock.
-func (gitUpdater *GitUpdater) observeWriteback(app string, d time.Duration) {
-	if gitUpdater.metrics != nil {
-		gitUpdater.metrics.ObserveGitWritebackDuration(app, d.Seconds())
-	}
-}
-
-// UpdateIfNeeded updates the git repository if the application is managed by the
-// watcher and has valid credentials. isSuperseded is an optional (at most one)
-// predicate forwarded to the write-back retry loop so a task superseded by a
-// newer deployment aborts instead of committing a stale image tag.
-func (gitUpdater *GitUpdater) UpdateIfNeeded(app *models.Application, task models.Task, isSuperseded ...func() bool) error {
-	if !app.IsManagedByWatcher() || !task.Validated {
-		slog.Debug("Skipping git repo update: application not managed by watcher or task not validated.", "id", task.Id)
-		return nil
-	}
-
-	gitopsRepo, err := models.NewGitopsRepo(app, gitUpdater.repoCachePath)
-	if err != nil {
-		slog.Error(fmt.Sprintf("Failed to get gitops repo info for app %s: %s", task.App, err), "id", task.Id)
-		return err
-	}
-
-	// Timed from just before the lock request so lock-wait captures the full queueing
-	// delay behind concurrent write-backs to the same repo, and the write-back timer
-	// captures only the work done once the lock is held.
-	lockRequested := time.Now()
-	gitUpdateFunc := func() error {
-		gitUpdater.observeLockWait(task.App, time.Since(lockRequested))
-		workStart := time.Now()
-		defer func() { gitUpdater.observeWriteback(task.App, time.Since(workStart)) }()
-		slog.Debug("Application managed by watcher. Initiating git repo update.", "id", task.Id)
-		return gitUpdater.updateGitRepo(app, &task, &gitopsRepo, isSuperseded...)
-	}
-
-	err = gitUpdater.locker.WithLock(gitopsRepo.RepoUrl, gitUpdateFunc)
-	if err != nil {
-		slog.Error(fmt.Sprintf("Failed git repo update for app %s: %s", task.App, err), "id", task.Id)
-		return err
-	}
-
-	return nil
-}
-
-func (gitUpdater *GitUpdater) updateGitRepo(app *models.Application, task *models.Task, gitopsRepo *models.GitopsRepo, isSuperseded ...func() bool) error {
-	// context.Background() is intentional: the call stack above this point
-	// does not carry a context. Propagating a cancellable context from
-	// WaitForRollout is a future improvement.
-	err := UpdateGitImageTag(context.Background(), app, task, gitopsRepo, updater.GitClient{}, isSuperseded...)
-	if err != nil {
-		slog.Error(fmt.Sprintf("Failed to update git repo. Error: %s", err.Error()), "id", task.Id)
-		return err
-	}
-	return nil
-}
 
 // ArgoStatusUpdater handles the monitoring and updating of ArgoCD application deployments
 type ArgoStatusUpdater struct {
@@ -552,47 +152,6 @@ func (updater *ArgoStatusUpdater) waitForApplicationDeployment(task models.Task)
 
 	// Wait for rollout completion
 	return updater.monitor.WaitRollout(task)
-}
-
-// handleApplicationFetchError ensures we don't retry for not found errors
-func handleApplicationFetchError(task models.Task, err error) error {
-	if task.IsAppNotFoundError(err) {
-		return retry.Unrecoverable(err)
-	}
-	slog.Debug(fmt.Sprintf("Failed fetching application status. Error: %s", err.Error()), "id", task.Id)
-	return err
-}
-
-// checkRolloutStatus checks if the application completed rollout successfully
-func checkRolloutStatus(task models.Task, application *models.Application, registryProxyUrl string, acceptSuspended bool) error {
-	status := application.GetRolloutStatus(task.ListImages(), registryProxyUrl, acceptSuspended)
-
-	switch status {
-	case models.ArgoRolloutAppDegraded:
-		slog.Debug("Application is degraded", "id", task.Id)
-		normalizedImages := helpers.NormalizeImages(application.Status.Summary.Images)
-		hash := helpers.GenerateHash(strings.Join(normalizedImages, ","))
-		if !bytes.Equal(task.SavedAppStatus.ImagesHash, hash) {
-			return retry.Unrecoverable(errors.New("application has degraded"))
-		}
-	case models.ArgoRolloutAppSuccess:
-		slog.Debug("Application rollout finished", "id", task.Id)
-		return nil
-	default:
-		slog.Debug(fmt.Sprintf("Application status is not final. Status received \"%s\"", status), "id", task.Id)
-	}
-	return errForceRetry
-}
-
-// determineFailureStatus converts API errors into appropriate status messages
-func determineFailureStatus(task models.Task, err error) string {
-	if task.IsAppNotFoundError(err) {
-		return models.StatusAppNotFoundMessage
-	}
-	if strings.Contains(err.Error(), argoUnavailableErrorMessage) {
-		return models.StatusAborted
-	}
-	return models.StatusFailedMessage
 }
 
 // sendNotification sends the task update through the configured notifier if available.
